Moved Household struct to its own file in order to fix circular file dependency due to Dist.h
